+++ composer.json
@@ -14,6 +14,7 @@
 		}
 	},
     "scripts": {
-        "lint": "phpcs --extensions=php -p"
-    }
+		"lint": "vendor/bin/phpcs --standard=./phpcs.xml --ignore=node_modules,vendor,build .",
+		"lint:fix": "vendor/bin/phpcbf --ignore=node_modules,vendor,build ."
+	}
 }
