+++ editor-styles.css
@@ -1,0 +1,1 @@
+@import "assets/css/editor-styles.css";
