+++ package.json
@@ -6,66 +6,38 @@
     "version": "3.3.0",
     "last_beta_version": "2.4.0-beta3",
     "scripts": {
+		"start": "wp-scripts start",
         "update-version": "node .github/scripts/update-version-in-files.js",
-        "build:prod": "grunt default --environment=production && WP_SRC_DIRECTORY=$(pwd) wp-scripts build --env=production",
-        "build:dev": "concurrently \"grunt default\" \"webpack --env=development\"",
-        "build:dev:watch": "concurrently \"grunt default watch\" \"webpack --env=developmentLocal\"",
+        "build:prod": "composer install --no-dev && wp-scripts build --env=production",
+        "build:dev": "composer update && composer install && wp-scripts build",
+        "build:dev:watch": "wp-scripts start",
         "clean:build": "rimraf build && rimraf hello-elementor",
         "zip": "npm run clean:build && npm run build:prod && rsync -av --exclude-from=.buildignore . hello-elementor && zip -r hello-elementor.$npm_package_version.zip hello-elementor/*",
         "lint:js": "eslint .",
         "build:new": "wp-scripts build"
     },
     "devDependencies": {
-        "@babel/core": "~7.18.0",
-        "@babel/plugin-proposal-class-properties": "^7.2.1",
-        "@babel/plugin-proposal-optional-chaining": "^7.9.0",
-        "@babel/plugin-syntax-dynamic-import": "^7.8.3",
-        "@babel/preset-env": "^7.12.11",
         "@wordpress/components": "^25.7.0",
         "@wordpress/eslint-plugin": "^12.1.0",
         "@wordpress/i18n": "^4.41.0",
         "@wordpress/notices": "^4.9.0",
 		"@wordpress/scripts": "^27.1.0",
         "autoprefixer": "~10.4.19",
-        "babel-eslint": "^8.2.6",
-        "babel-loader": "8.0.0",
-        "concurrently": "^6.2.0",
 		"copy-webpack-plugin": "^12.0.2",
-        "cssnano": "4.1.11",
-        "eslint": "8.36.0",
-        "eslint-config-wordpress": "^2.0.0",
-        "eslint-plugin-babel": "^5.3.1",
-        "eslint-plugin-jsx-a11y": "^6.5.1",
-        "eslint-plugin-no-jquery": "^2.7.0",
-        "eslint-plugin-react": "^7.29.4",
-        "eslint-webpack-plugin": "^4.0.0",
-        "grunt": "^1.4.1",
-        "grunt-checktextdomain": "^1.0.1",
-        "grunt-contrib-sass": "~1.0.0",
-        "grunt-contrib-watch": "~1.1.0",
-        "grunt-sass": "~3.1.0",
-        "grunt-text-replace": "~0.4.0",
-        "grunt-webpack": "^6.0.0",
-        "grunt-wp-readme-to-markdown-with-extra": "~2.2.0",
-        "load-grunt-tasks": "^4.0.0",
+		"eslint-plugin-babel": "^5.3.1",
+		"eslint-plugin-no-jquery": "^3.0.2",
+		"eslint-plugin-react": "^7.35.2",
         "matchdep": "~2.0.0",
         "npm-build-zip": "^1.0.3",
         "rimraf": "^3.0.2",
-        "sass": "^1.53.0",
-        "terser-webpack-plugin": "^5.0.3",
-        "webpack": "^5.11.0",
 		"webpack-cli": "^5.1.4",
 		"webpack-remove-empty-scripts": "^1.0.4"
 	},
     "dependencies": {
         "@babel/plugin-transform-modules-commonjs": "^7.2.0",
         "@babel/plugin-transform-runtime": "^7.3.4",
-        "@elementor/eslint-plugin": "^1.0.0",
         "@elementor/icons": "1.37.0",
         "@elementor/ui": "1.20.3",
-        "@lodder/grunt-postcss": "^3.1.1",
-        "grunt-autoprefixer": "^3.0.4",
-        "grunt-cssnano": "^2.1.0",
         "html-entities": "^2.5.2",
         "react": "^18.3.1",
         "react-dom": "^18.3.1"
