+++ theme.css
@@ -1,0 +1,1 @@
+@import "assets/css/theme.css";
