+++ webpack.config.js
@@ -1,14 +1,26 @@
-/**
- * Grunt webpack task config
- *
- * @package
- */
-const path = require( 'path' );
+// WordPress webpack config.
 const defaultConfig = require( '@wordpress/scripts/config/webpack.config' );
 
-const TerserPlugin = require( 'terser-webpack-plugin' );
+// Plugins.
+const RemoveEmptyScriptsPlugin = require( 'webpack-remove-empty-scripts' );
+const CopyWebpackPlugin = require( 'copy-webpack-plugin' );
 
-const entry = {
+// Utilities.
+const path = require( 'path' );
+const imagesPath = path.resolve( __dirname, './build/images' );
+
+const modulesDir = process.cwd() + '/modules/';
+const scssPath = path.resolve( __dirname, './assets/scss' );
+
+const entryPoints = {
+
+	'../css/reset': path.resolve( scssPath, 'reset.scss' ),
+	'../css/theme': path.resolve( scssPath, 'theme.scss' ),
+	'../css/header-footer': path.resolve( scssPath, 'header-footer.scss' ),
+	'../css/editor': path.resolve( scssPath, 'editor.scss' ),
+	'../css/editor-styles': path.resolve( scssPath, 'editor-styles.scss' ),
+	'../css/customizer': path.resolve( scssPath, 'customizer.scss' ),
+
 	'hello-editor': path.resolve( __dirname, './assets/dev/js/editor/hello-editor.js' ),
 	'hello-frontend': path.resolve( __dirname, './assets/dev/js/frontend/hello-frontend.js' ),
 	'hello-home-app': path.resolve( __dirname, './modules/admin-home/assets/js//hello-elementor-admin.js' ),
@@ -18,94 +30,23 @@ const entry = {
 	'hello-conversion-banner': path.resolve( __dirname, './modules/admin-home/assets/js//hello-elementor-conversion-banner.js' ),
 };
 
-const moduleRules = {
-	rules: [
-		...defaultConfig.module.rules,
-		{
-			test: /\.js$/,
-			exclude: /node_modules/,
-			use: [
-				{
-					loader: 'babel-loader',
-					options: {
-						presets: [ '@babel/preset-env', '@babel/preset-react' ],
-						plugins: [
-							[ '@babel/plugin-proposal-class-properties' ],
-							[ '@babel/plugin-transform-runtime' ],
-							[ '@babel/plugin-transform-modules-commonjs' ],
-							[ '@babel/plugin-proposal-optional-chaining' ],
-							[ '@babel/plugin-syntax-dynamic-import' ],
-						],
-					},
-				},
-			],
-		},
-	],
-};
-
-const commonConfig = {
+module.exports = {
 	...defaultConfig,
-	target: 'web',
-	context: __dirname,
-	module: moduleRules,
-	entry,
-	output: {
-		...defaultConfig.output,
-		path: path.resolve( __dirname, './assets/js' ),
-		filename: '[name].js',
-	},
-};
+	...{
+		entry: entryPoints,
+		output: {
+			...defaultConfig.output,
+			path: path.resolve( __dirname, './assets/js' ),
+		},
+		plugins: [
+			// Include WP's plugin config.
+			...defaultConfig.plugins,
 
-const webpackConfig = {
-	...commonConfig,
-	mode: 'development',
-	output: {
-		...commonConfig.output,
-		devtoolModuleFilenameTemplate: './[resource]',
-	},
-	entry: {
-		...entry,
-	},
-	devtool: 'source-map',
-};
-
-const webpackProductionConfig = {
-	...commonConfig,
-	mode: 'production',
-	optimization: {
-		...defaultConfig.optimization || {},
-		minimize: true,
-		minimizer: [
-			new TerserPlugin( {
-				terserOptions: {
-					keep_fnames: true,
-				},
-				include: /\.min\.js$/,
+			// Removes the empty `.js` files generated by webpack but
+			// sets it after WP has generated its `*.asset.php` file.
+			new RemoveEmptyScriptsPlugin( {
+				stage: RemoveEmptyScriptsPlugin.STAGE_AFTER_PROCESS_PLUGINS,
 			} ),
 		],
 	},
-	performance: { hints: false },
-};
-
-// Add minified entry points
-Object.entries( webpackProductionConfig.entry ).forEach( ( [ wpEntry, value ] ) => {
-	webpackProductionConfig.entry[ wpEntry + '.min' ] = value;
-} );
-
-webpackProductionConfig.plugins = defaultConfig.plugins;
-
-module.exports = ( env ) => {
-	if ( env.developmentLocal ) {
-		return { ...webpackConfig, watch: true };
-	}
-
-	if ( env.production ) {
-		return webpackProductionConfig;
-	}
-
-	if ( env.development ) {
-		return webpackConfig;
-	}
-
-	throw new Error( 'missing or invalid --env= development/production/developmentWithWatch/productionWithWatch' );
 };
