--- webpack.config.new.js
@@ -1,66 +1,0 @@
-// WordPress webpack config.
-const defaultConfig = require( '@wordpress/scripts/config/webpack.config' );
-
-// Plugins.
-const RemoveEmptyScriptsPlugin = require( 'webpack-remove-empty-scripts' );
-const CopyWebpackPlugin = require( 'copy-webpack-plugin' );
-
-// Utilities.
-const path = require( 'path' );
-const imagesPath = path.resolve( __dirname, './build/images' );
-
-const modulesDir = process.cwd() + '/modules/';
-const scssPath = path.resolve( __dirname, './assets/scss' );
-
-const entryPoints = {
-
-	// Theme module
-	'css/style': path.resolve( scssPath, 'theme.scss' ),
-	'css/theme': path.resolve( scssPath, 'theme.scss' ),
-	'css/header-footer': path.resolve( scssPath, 'header-footer.scss' ),
-	'css/editor': path.resolve( scssPath, 'editor.scss' ),
-	'css/editor-styles': path.resolve( scssPath, 'editor-styles.scss' ),
-	'css/customizer': path.resolve( scssPath, 'customizer.scss' ),
-
-	'hello-editor': path.resolve( __dirname, './assets/dev/js/editor/hello-editor.js' ),
-	'hello-frontend': path.resolve( __dirname, './assets/dev/js/frontend/hello-frontend.js' ),
-	'hello-admin': path.resolve( __dirname, './assets/dev/js/admin/hello-admin.js' ),
-};
-
-module.exports = {
-	...defaultConfig,
-	...{
-		entry: entryPoints,
-		output: {
-			...defaultConfig.output,
-			path: path.resolve( __dirname, './build' ),
-		},
-		plugins: [
-			// Include WP's plugin config.
-			...defaultConfig.plugins,
-
-/*			new CopyWebpackPlugin( {
-				patterns: [
-					{
-						from: path.resolve( modulesDir, 'admin-home/assets/images' ),
-						to: imagesPath,
-					},
-					{
-						from: path.resolve( modulesDir, 'conversion-banner/assets/images' ),
-						to: imagesPath,
-					},
-					{
-						from: path.resolve( modulesDir, 'theme/assets/images' ),
-						to: imagesPath,
-					},
-				],
-			} ),*/
-
-			// Removes the empty `.js` files generated by webpack but
-			// sets it after WP has generated its `*.asset.php` file.
-			new RemoveEmptyScriptsPlugin( {
-				stage: RemoveEmptyScriptsPlugin.STAGE_AFTER_PROCESS_PLUGINS,
-			} ),
-		],
-	},
-};
